|<- -    Added some model tests with rspec    -->|
This changed has been made in order to test the model, also I deleted the column picture from the user table because it was not necessary

This changed was accomplished by creating the spec files for each model and writing the tests there. For the column I generated a migration to delete it.